refactor: (#75) game state 인터페이스 통합 및 InMemory 스토어 제거
