+++ src/game/game.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TopicGateway } from './topic/topic.gateway';
 import { TopicService } from './topic/topic.service';
-import { GAME_STATE_STORE } from './interfaces/game-state-store.interfaces';
-import { InMemoryGameStateStore } from './infra/game-state.store.inmemory';
 import { GameGateway } from './game.gateway';
 import { ChatModule } from 'src/chat/chat.module';
 import { JwtModule } from '@nestjs/jwt';
@@ -11,6 +9,7 @@ import { GameSessionService } from './session/game-session.service';
 import { WaitingModule } from 'src/waiting/waiting.module';
 import { GameStateStore } from 'src/game-state/game-state.store';
 import { RedisModule } from 'src/redis/redis.module';
+import { GAME_STATE_STORE } from 'src/game-state/interfaces/game-state.interface';
 
 @Module({
   imports: [
