+++ src/game/session/game-session.service.ts
@@ -1,13 +1,16 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { GAME_STATE_STORE, GamePhase } from '../interfaces/game-state-store.interfaces';
 import { GAME_EVENT_PUBLISHER } from '../interfaces/game-event-publisher.interfaces';
-import type { IGameStateStore } from '../interfaces/game-state-store.interfaces';
 import type { IGameEventPublisher } from '../interfaces/game-event-publisher.interfaces';
 import { TopicService } from '../topic/topic.service';
 import { GameSessionEntity } from '../entities/game-session.entity';
 import { RANDOM_THEMES } from '../topic/constants/topic.constant';
 import { GAME_ERRORS, GAME_EVENTS } from '../constants/game.constant';
 import { OnEvent } from '@nestjs/event-emitter';
+import {
+  GAME_STATE_STORE,
+  GamePhase,
+  type IGameStateStore,
+} from 'src/game-state/interfaces/game-state.interface';
 
 @Injectable()
 export class GameSessionService {
